requirements.txt
# Core data processing
pyspark==3.5.1
delta-spark==3.2.0
pandas==2.2.2
pyarrow==16.1.0
numpy==1.26.4

# Kafka / streaming
confluent-kafka==2.4.0
faker==25.8.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==1.4.52

# Airflow
apache-airflow==2.9.3
apache-airflow-providers-postgres==5.11.1
apache-airflow-providers-apache-spark==4.8.1
apache-airflow-providers-common-sql==1.14.1

# Quality and config
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
structlog==24.2.0
tenacity==8.5.0

# Testing
pytest==8.2.2
pytest-mock==3.14.0